Allow only one Merlin version
Before, [merl-an] could be fed different Merlin versions or Merlin
frontends (i.e. [single] or [server]) and [merl-an] would run all queries
for all fed Merlin versions/frontends. That can be useful for the
[performance] cmd. However, it's unnecessary noise for other commands
and makes the code harder to refactor.

This commit removes that feature. With it, [merl-an] runs the query on
exactly one Merlin version and frontend.